[ROCm] fix for multiple runners and docker commands (#79759)
The ROCm GHA workflow was initially based on the assumption that only a single runner exists on the host.  Now that ROCm CI hosts sometimes have two runners splitting the available GPUs, the docker commands to clean up existing containers and stale images cannot assume a single runner instance.  This updates the docker commands to only stop the current container and to safely ignore when a docker prune is already in progress.

CC @jithunnair-amd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79759
Approved by: https://github.com/janeyx99